Add BUILD_SWIFT environment switch to build
Sometimes it can be hard to read the errors generated by swift-bridge.
In these cases, it can be better to allow the rust compiler to give more
useful errors. Change the default build configuration to not try to generate
the swift FFI code. When you're ready to generate the code, set BUILD_SWIFT
environment variable to 1.